+++ build-deb.sh
@@ -2,7 +2,7 @@
 
 # Build the project
 echo "Building sysreport..."
-VERSION="0.5.0"
+VERSION="0.6.0"
 make clean
 make